Fix bug in Tree object writer that would cause lines to be dropped.
This would happen when the buffer to which we were writing was too small
to fit the new line.  In that case, we would double the size of the
buffer to make room, but we would not try to write the line down again.

This would've been caught by existing inline tests, a number of buffer
initialization methods are meant to use a length of 1 during inline tests,
however it appears that Core.am_running_inline_test does not work correctly
at this point.  I've not tried to fix this at this point.