Fix exception being raised in zlib finalizer.
inflateEnd/deflateEnd can return Z_DATA_ERROR when called on a
stream that was left in an intermediate state.  This can happen
when inflate/deflate was never called with the Z_FINISH flush
parameter.  Regardless, all of the allocated memory is still
released correctly after the call, so we should be happy to
tolerate this error and not raise.

Additionally, fix the exception raising code to properly relay
the return code of the method that failed.