[Object_store.size] always waits to receive a [size].
When reading from a non-packed file, [push_back] would always return
[`Reader_close] because we're not interested in receiving any file
contents.  There's a potential, yet extremely unlikely bug that can
happen in which [push_back] could be called before [on_header] if
the first chunk of compressed zlib data produces no output for the
raw object parser.

This feels incredibly unlikely with the current object reader code,
but a future commit is changing the object reader so that it
performs an initial small read before reading regular sized chunks
to try and obtain the header without reading unnecessary data in
the case of readers such as [size] that only care about the header.